requirements.txt
Flask
python-telegram-bot
Pillow
gspread
google-auth
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
